concatenated ann.py to model.py
